requirements.txt
flask
scapy
python-nmap
netaddr
requests
jinja2 